added a basic player script
Made it so player has a player position, and an empty move method for now which takes in a tile for the endposition.

The snake Move Player method will check if the head has a current player, and so i added a .hasPlayer() method in the Tiles, and if it does have a player on the head, return that player (through player get method) and move it to the tail.